\name{quantregRanger-package}
\alias{quantregRanger-package}
\docType{package}
\title{
\packageTitle{quantregRanger}
}
\description{
\packageDescription{quantregRanger}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{quantregRanger}
\packageIndices{quantregRanger}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{quantregRanger}

Maintainer: \packageMaintainer{quantregRanger}
}
\references{
N. Meinshausen (2006)
"Quantile Regression Forests", Journal of Machine Learning Research 7, 983-999
\url{http://jmlr.csail.mit.edu/papers/v7/}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
