% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.quantregRanger}
\alias{predict.quantregRanger}
\title{quantregRanger prediction}
\usage{
\method{predict}{quantregRanger}(object, data = NULL, quantiles = c(0.1,
  0.5, 0.9))
}
\arguments{
\item{object}{\code{quantregRanger} object.}

\item{data}{New test data of class \code{data.frame}}

\item{quantiles}{Numeric vector of quantiles that should be estimated}
}
\value{
A matrix. The first column contains the conditional quantile 
estimates for the first entry in the vector quantiles. The second 
column contains the estimates for the second entry of quantiles and so on.
}

