% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantregRanger.R
\name{quantregRanger}
\alias{quantregRanger}
\title{Quantile Regression with Ranger

Creates a quantile regression forest like described in Meinshausen, 2006.##'}
\usage{
quantregRanger(formula = NULL, data = NULL, num.trees = 500,
  mtry = NULL, min.node.size = NULL)
}
\arguments{
\item{formula}{Object of class \code{formula} or \code{character} describing the model to fit.}

\item{data}{Training data of class \code{data.frame}, \code{matrix} or \code{gwaa.data} (GenABEL).}

\item{num.trees}{Number of trees.}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables.}

\item{min.node.size}{Minimal node size. Default is 10.}
}
\examples{
y = rnorm(150)
x = cbind(y + rnorm(150), rnorm(150))
data = data.frame(x,y)
mod = quantregRanger(y ~ ., data = data)
predict(mod, data = data[1:5, ])

# QuantregForest
library(quantregForest)
mod2 = quantregForest(x=as.matrix(x), y=y)
predict(mod2, newdata = as.matrix(x[1:5,]), all=TRUE)

}

