\name{quantregRanger-package}
\alias{quantregRanger-package}
\docType{package}
\title{
\packageTitle{quantregRanger}
}
\description{
\packageDescription{quantregRanger}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{quantregRanger}
\packageIndices{quantregRanger}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{quantregRanger}

Maintainer: \packageMaintainer{quantregRanger}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
