% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.quantregRanger}
\alias{predict.quantregRanger}
\title{quantregRanger prediction}
\usage{
\method{predict}{quantregRanger}(object, data = NULL, quantiles = c(0.1,
  0.5, 0.9), all = TRUE, obs = 1, ...)
}
\arguments{
\item{object}{\code{quantregRanger} object.}

\item{data}{New test data of class \code{data.frame}}

\item{quantiles}{Numeric vector of quantiles that should be estimated}

\item{all}{A logical value. all=TRUE uses all observations for prediction. 
all=FALSE uses only a certain number of observations per node for prediction 
(set with argument obs). The default is all=TRUE}

\item{obs}{An integer number. Determines the maximal number of observations per node}

\item{...}{Currently ignored.
to use for prediction. The input is ignored for all=TRUE. The default is obs=1}
}
\value{
A matrix. The first column contains the conditional quantile 
estimates for the first entry in the vector quantiles. The second 
column contains the estimates for the second entry of quantiles and so on.
}
\description{
Predicts quantiles for a quantile regression forest trained with quantregRanger.
}
